The transition to Kivy in Progress...
Hello everyone,

So we start this project with the Tkinter UI but at the begining of the project , kivy was already there in my mind.

I wanted to learn more and its not finish ... it's just the begining for sure .

It's like a challenge for me because it's kind of hard to dev (to me) with OOP (Objetc-Oriented Programming) way. When I was at school , it's there that I learn how to dev and it's was hard for me to understand the logic under the OOP way but I want to face to my weakness and become better to get stronger skills for programming.

That's why I'm working hard on this project Tkinter + Kivy.